Update README.
Update README.

Update README.

Throw errors were possible.

Sometimes racr is simply empty. If it is empty HVAC is not on (an assumption).

Change homepage to Gitlab.

For some odd reason VehicleInfoList is not present on 1th gen Leafs. It is only there for 2nd gen Leafs. Should fix problems for people with 1th gen Leaf.

For some reason electricCostScale can vary from country to country.

.